+++ components/Layout/SideBar.tsx
@@ -8,15 +8,19 @@ import {
     StyledAboutBody,
     StyledContainerTool,
     StyledImageGit,
-} from './styles'
-import { Contact } from './Contact'
-import type { HeadProps } from './interface'
-import { About } from './About'
-import { Language } from './Language'
-import { Lib } from './Lib'
-import { Skill } from './Skill'
+    StyledZoomIcon,
+    StyledWrapperZoomImage,
+    StyledContainerOpacity,
+    StyledContainerZoomImage,
+} from './style'
+import { Contact } from '@containers/Home/Contact'
+import type { HeadProps } from '@containers/Home/interface'
+import { About } from '@containers/Home/About'
+import { Language } from '@containers/Home/Language'
+import { Lib } from '@containers/Home/Lib'
+import { Skill } from '@containers/Home/Skill'
 import { useIsPrintMode } from 'hooks'
-import { useContext } from 'react'
+import { useContext, useState } from 'react'
 import { ThemeContext } from '@context/ThemeContext'
 import { Theme } from 'style/Theme'
 
@@ -29,12 +33,16 @@ const handlePrint = () => {
 export const SideBar: React.FC<HeadProps> = ({ data }) => {
     const { theme } = useContext(ThemeContext)
     const currentTheme = Theme[theme]
-
     const isPrintMode = useIsPrintMode()
+
+    const [isZoomIcon, setIsZoomIcon] = useState(false)
+    const resizeImg = () => {
+        setIsZoomIcon(!isZoomIcon)
+    }
     return (
         <>
             <StyledAboutHead>
-                <StyledWrapperImage>
+                <StyledWrapperImage onClick={resizeImg}>
                     <StyledIcon src="/logo.jpg" layout="fill" objectFit="cover" />
                 </StyledWrapperImage>
                 <StyledName>Luu The Nguyen</StyledName>
@@ -73,6 +81,23 @@ export const SideBar: React.FC<HeadProps> = ({ data }) => {
             <StyledAboutFooter>
                 <Contact contact={data.contacts} />
             </StyledAboutFooter>
+            {isZoomIcon && (
+                <>
+                    <StyledContainerOpacity onClick={resizeImg} />
+                    <StyledContainerZoomImage onClick={resizeImg}>
+                        <StyledWrapperZoomImage>
+                            <StyledZoomIcon
+                                src="/logo.jpg"
+                                layout="responsive"
+                                objectFit="cover"
+                                width="30vw"
+                                height="30vh"
+                                sizes="32w"
+                            />
+                        </StyledWrapperZoomImage>
+                    </StyledContainerZoomImage>
+                </>
+            )}
         </>
     )
 }
