+++ components/Layout/interface.ts
@@ -1,4 +1,6 @@
+import { DataHeadProps } from '@containers/Home'
 export interface Props {
     name: string
     version: string
+    headData: DataHeadProps
 }
